"Updated normalize.css: added appearance and margin styles for number input type and search pseudo-class"
This commit message describes the changes in the code diff for the `normalize.css` file. It mentions the addition of new styles for the `input[type="number"]` element, which include setting the `-moz-appearance` property to `textfield` and adding styles for the `webkit-inner-spin-button` and `webkit-outer-spin-button` pseudo-elements. It also mentions the addition of styles for the `[type="search"]` pseudo-class. The message is concise and accurately describes the changes made in the code diff.